requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
conda==4.3.16
distlib==0.3.1
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pipenv==2020.8.13
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==1.0.1
